+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fix date input where typing the age with keyboard was resetting caret position.
 - Fix link to GitHub in settings.
 
 ## [1.0.0]